+++ proto/src/prost/google.api.rs
@@ -7,7 +7,7 @@ pub struct Http {
     ///
     /// **NOTE:** All service configuration rules follow "last one wins" order.
     #[prost(message, repeated, tag="1")]
-    pub rules: ::std::vec::Vec<HttpRule>,
+    pub rules: ::prost::alloc::vec::Vec<HttpRule>,
     /// When set to true, URL path parmeters will be fully URI-decoded except in
     /// cases of single segment matches in reserved expansion, where "%2F" will be
     /// left encoded.
@@ -240,29 +240,30 @@ pub struct HttpRule {
     ///
     /// Refer to [selector][google.api.DocumentationRule.selector] for syntax details.
     #[prost(string, tag="1")]
-    pub selector: std::string::String,
+    pub selector: ::prost::alloc::string::String,
     /// The name of the request field whose value is mapped to the HTTP body, or
     /// `*` for mapping all fields not captured by the path pattern to the HTTP
     /// body. NOTE: the referred field must not be a repeated field and must be
     /// present at the top-level of request message type.
     #[prost(string, tag="7")]
-    pub body: std::string::String,
+    pub body: ::prost::alloc::string::String,
     /// Optional. The name of the response field whose value is mapped to the HTTP
     /// body of response. Other response fields are ignored. When
     /// not set, the response message will be used as HTTP body of response.
     #[prost(string, tag="12")]
-    pub response_body: std::string::String,
+    pub response_body: ::prost::alloc::string::String,
     /// Additional HTTP bindings for the selector. Nested bindings must
     /// not contain an `additional_bindings` field themselves (that is,
     /// the nesting may only be one level deep).
     #[prost(message, repeated, tag="11")]
-    pub additional_bindings: ::std::vec::Vec<HttpRule>,
+    pub additional_bindings: ::prost::alloc::vec::Vec<HttpRule>,
     /// Determines the URL pattern is matched by this rules. This pattern can be
     /// used with any of the {get|put|post|delete|patch} methods. A custom method
     /// can be defined using the 'custom' field.
     #[prost(oneof="http_rule::Pattern", tags="2, 3, 4, 5, 6, 8")]
-    pub pattern: ::std::option::Option<http_rule::Pattern>,
+    pub pattern: ::core::option::Option<http_rule::Pattern>,
 }
+/// Nested message and enum types in `HttpRule`.
 pub mod http_rule {
     /// Determines the URL pattern is matched by this rules. This pattern can be
     /// used with any of the {get|put|post|delete|patch} methods. A custom method
@@ -271,19 +272,19 @@ pub mod http_rule {
     pub enum Pattern {
         /// Used for listing and getting information about resources.
         #[prost(string, tag="2")]
-        Get(std::string::String),
+        Get(::prost::alloc::string::String),
         /// Used for updating a resource.
         #[prost(string, tag="3")]
-        Put(std::string::String),
+        Put(::prost::alloc::string::String),
         /// Used for creating a resource.
         #[prost(string, tag="4")]
-        Post(std::string::String),
+        Post(::prost::alloc::string::String),
         /// Used for deleting a resource.
         #[prost(string, tag="5")]
-        Delete(std::string::String),
+        Delete(::prost::alloc::string::String),
         /// Used for updating a resource.
         #[prost(string, tag="6")]
-        Patch(std::string::String),
+        Patch(::prost::alloc::string::String),
         /// The custom pattern is used for specifying an HTTP method that is not
         /// included in the `pattern` field, such as HEAD, or "*" to leave the
         /// HTTP method unspecified for this rule. The wild-card rule is useful
@@ -297,8 +298,8 @@ pub mod http_rule {
 pub struct CustomHttpPattern {
     /// The name of this custom HTTP verb.
     #[prost(string, tag="1")]
-    pub kind: std::string::String,
+    pub kind: ::prost::alloc::string::String,
     /// The path matched by this custom verb.
     #[prost(string, tag="2")]
-    pub path: std::string::String,
+    pub path: ::prost::alloc::string::String,
 }