+++ CHANGELOG.md
@@ -20,11 +20,13 @@
 - [ibc-relayer-cli]
   - Added `create connection` and `create channel` CLIs ([#630], [#715])
   - Proposed ADR 006 to describe Hermes v0.2.0 use-cases ([#637])
+  - Update gaia to version 4.1.0 for e2e tests on CI ([#702])
 
 ### IMPROVEMENTS
 
 - [ibc]
   - Follow Rust guidelines naming conventions ([#689])
+  - Per client structure modules ([#740])
 
 - [ibc-relayer]
   - [nothing yet]
@@ -66,6 +68,8 @@
 [#699]: https://github.com/informalsystems/ibc-rs/issues/699
 [#700]: https://github.com/informalsystems/ibc-rs/pull/700
 [#715]: https://github.com/informalsystems/ibc-rs/issues/715
+[#702]: https://github.com/informalsystems/ibc-rs/issues/702
+[#740]: https://github.com/informalsystems/ibc-rs/issues/740
 
 ## v0.1.1
 *February 17, 2021*